+++ app/Filament/Resources/DomainTlds/Pages/EditDomainTld.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace App\Filament\Resources\DomainTlds\Pages;
+
+use App\Filament\Resources\DomainTlds\DomainTldResource;
+use Filament\Resources\Pages\EditRecord;
+
+class EditDomainTld extends EditRecord
+{
+    protected static string $resource = DomainTldResource::class;
+}